food11/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
fastapi
uvicorn[standard]
torch==2.5.1+cpu
torchvision==0.20.1+cpu
pillow
numpy
